feat(vats): makeMockChainStorageRoot test util
